More code review comment fixes for ealier checkins.
